move github oauth logic into package
